+++ lib/actions/settings/shopSyrattachPluginSettings.action.php
@@ -20,7 +20,7 @@ class shopSyrattachPluginSettingsAction extends waViewAction
     public function execute()
     {
         if (!$this->getUser()->getRights('shop', 'settings')) {
-            throw new waException(_w('Access denied'));
+            throw new waException(_wp('Access denied'));
         }
 
         $plugin = wa('shop')->getPlugin('syrattach');